requirements.txt
pymongo[srv]==4.2.0
scikit-learn
pandas
numpy
pyYAML
scipy
ipykernel
imblearn
xgboost
dill
-e .